requirements.txt
selenium==4.5.0
python-dotenv==0.21.0
